Exception Handling - Delete

+++ src/main/java/com/muhdo/course/services/UserService.java
@@ -2,8 +2,10 @@ package com.muhdo.course.services;
 
 import com.muhdo.course.entities.User;
 import com.muhdo.course.repositories.UserRepository;
+import com.muhdo.course.services.exceptions.DatabaseException;
 import com.muhdo.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +30,15 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
